+++ src/internal/OpenApiTools/components/RequestBody.ts
@@ -1,4 +1,4 @@
-import * as ts from "typescript";
+import ts from "typescript";
 
 import type { OpenApi } from "../../../types";
 import { Factory } from "../../TsGenerator";
@@ -42,10 +42,11 @@ export const generateNamespace = (
   context: ToTypeNode.Context,
   converterContext: ConverterContext.Types,
 ): void => {
+  const escapeName = converterContext.escapeDeclarationText(name);
   const basePath = `${parentName}/${name}`;
   store.addStatement(basePath, {
     kind: "namespace",
-    name,
+    name: converterContext.escapeDeclarationText(escapeName),
     comment: requestBody.description,
   });
   store.addStatement(`${basePath}/Content`, {